// src/app/models/post.ts
export class Post{
    id?: any;
    content?: string;
    medias?: { media1?: any, media2?: any, media3?: any };
    date?: any; // Date timestamp
    createAt?: Date; // created date in date format
    uid?: any;
    displayName?: any;
    photoURL?: any;
    categoryId?: any;
    likes?: any[];
    comments?: any[];
}